+++ ci-jobs/master-build.yml
@@ -1,7 +1,4 @@
-# Node.js
-# Build a general Node.js application with npm.
-# Add steps that analyze code, save build artifacts, deploy, and more:
-# https://docs.microsoft.com/vsts/pipelines/languages/javascript
+# Master build. Runs on any commit to master.
 jobs:
 - job: crowdin
   steps:
