+++ ci-jobs/package.yml
@@ -1,7 +1,4 @@
-# Node.js
-# Build a general Node.js application with npm.
-# Add steps that analyze code, save build artifacts, deploy, and more:
-# https://docs.microsoft.com/vsts/pipelines/languages/javascript
+# Packager. Runs on refs/tags/*
 jobs:
 - template: templates/package.yml
   parameters:
