+++ ci-jobs/pr-validation.yml
@@ -1,7 +1,4 @@
-# Node.js
-# Build a general Node.js application with npm.
-# Add steps that analyze code, save build artifacts, deploy, and more:
-# https://docs.microsoft.com/vsts/pipelines/languages/javascript
+# Runs on pull requests
 jobs:
 - template: templates/build.yml
   parameters:
