+++ ci-jobs/templates/build.yml
@@ -28,8 +28,8 @@ jobs:
       displayName: NPM test
     - script: npx build --publish never --dir
       displayName: Build application bundles
-    #- script: npm run e2e
-      #displayName: Run end-to-end tests
+    - script: npm run test-e2e
+      displayName: Run end-to-end tests
     - task: PublishTestResults@2
       displayName: 'Publish Test Results'
       inputs:
