+++ package.json
@@ -30,11 +30,13 @@
     "build": "run-s build:*",
     "build:main": "cross-env NODE_ENV=production parcel build --target node --public-url . --out-dir ./dist --out-file main.js ./app/main.js",
     "build:renderer": "cross-env NODE_ENV=production parcel build --no-content-hash --out-dir ./dist --target electron --public-url . ./app/renderer/index.html",
-    "watch": "concurrently \"npm run watch:main\" \"npm run watch:renderer\"",
+    "watch": "concurrently \"npm run watch-main\" \"npm run watch-renderer\"",
     "watch:main": "cross-env NODE_ENV=development parcel --target node --out-dir ./dist --out-file main.js ./app/main.js",
     "watch:renderer": "cross-env NODE_ENV=development parcel --out-dir ./dist --target electron --public-url . ./app/renderer/index.html",
     "start": "cross-env RUNNING_LOCALLY=true FORCE_NO_WRONG_FOLDER=true electron ./",
     "clean": "rimraf dist/",
+    "watch-main": "cross-env NODE_ENV=development parcel --target node --out-dir ./dist --out-file main.js ./app/main.js",
+    "watch-renderer": "cross-env NODE_ENV=development parcel --out-dir ./dist --target electron --public-url . ./app/renderer/index.html",
     "build-main-dev": "cross-env NODE_ENV=development parcel build --target node --out-dir ./dist --out-file main.js ./app/main.js",
     "build-renderer-dev": "cross-env NODE_ENV=development parcel build --out-dir ./dist --target electron --public-url . ./app/renderer/index.html",
     "start-dev": "cross-env RUNNING_LOCALLY=true FORCE_NO_WRONG_FOLDER=true NODE_ENV=development electron --inspect ./",
