+++ test/e2e/main-e2e.test.js
@@ -1,5 +1,4 @@
-import { Application } from 'spectron';
-import { fs, logger } from 'appium-support';
+import { fs } from 'appium-support';
 import B from 'bluebird';
 import os from 'os';
 import path from 'path';
@@ -9,55 +8,10 @@ import dirCompare from 'dir-compare';
 import { retryInterval } from 'asyncbox';
 import MainPage from './pages/main-page-object';
 
-const log = logger.getLogger('E2E Test');
 const platform = os.platform();
 
 chai.should();
 chai.use(chaiAsPromised);
-
-let appPath;
-let args = [];
-if (process.env.SPECTRON_USE_APPIUM_BINARY) {
-  if (platform === 'linux') {
-    appPath = path.join(__dirname, '..', '..', 'release', 'linux-unpacked', 'appium-desktop');
-  } else if (platform === 'darwin') {
-    appPath = path.join(__dirname, '..', '..', 'release', 'mac', 'Appium.app', 'Contents', 'MacOS', 'Appium');
-  } else if (platform === 'win32') {
-    appPath = path.join(__dirname, '..', '..', 'release', 'win-ia32-unpacked', 'Appium.exe');
-  }
-} else {
-  appPath = require('electron');
-  args.push(path.join(__dirname, '..', '..'));
-}
-
-before(async function () {
-  this.timeout(process.env.E2E_TIMEOUT || 60 * 1000);
-  log.info(`Running Appium from: ${appPath}`);
-  log.info(`Checking that "${appPath}" exists`);
-  const applicationExists = await fs.exists(appPath);
-  if (!applicationExists) {
-    log.error(`Could not run tests. "${appPath}" does not exist.`);
-    process.exit(1);
-  }
-  log.info(`App exists. Creating Spectron Application instance`);
-  this.app = new Application({
-    path: appPath,
-    args,
-    env: {
-      FORCE_NO_WRONG_FOLDER: true,
-      SKIP_AUTO_UPDATE: true,
-    }
-  });
-  log.info(`Spectron Application instance created. Starting app`);
-  await this.app.start();
-  log.info(`App started`);
-});
-
-after(function () {
-  if (this.app && this.app.isRunning()) {
-    return this.app.stop();
-  }
-});
 
 describe('application launch', function () {
 
@@ -89,7 +43,7 @@ describe('application launch', function () {
   it('check that WebDriverAgent folder is the same in /releases as it is in /node_modules (regression test for https://github.com/appium/appium-desktop/issues/417)', async function () {
     // NOTE: This isn't really an "e2e" test, but the test has to be written here because the /release
     // folder needs to be built in order to run the test
-    if (platform !== 'darwin') {
+    if (platform !== 'darwin' || !process.env.SPECTRON_TEST_PROD_BINARIES) {
       return this.skip();
     }
